Add files via upload
Lung nodule classification is the process of categorizing lung nodules detected in medical imaging into different classes, such as benign or malignant. It involves analyzing the characteristics of the nodules, such as size, shape, density, and texture, to determine the likelihood of malignancy. Machine learning techniques, such as deep learning algorithms, are commonly used to aid in the automated classification of lung nodules based on these features.